Move start to own namespace
